[Harness] Split out classes from Jenkins.cs to separate files (#8046)
