+++ tests/xharness/AppRunner.cs
@@ -3,15 +3,13 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using System.Xml.Xsl;
 using Xamarin;
-using Xamarin.Utils;
+using Xharness.Jenkins.TestTasks;
 using Xharness.Listeners;
 using Xharness.Logging;
