+++ tests/xharness/Harness.cs
@@ -685,7 +685,7 @@ namespace Xharness
 				AutoConfigureMac (false);
 			}
 			
-			var jenkins = new Jenkins ()
+			var jenkins = new Jenkins.Jenkins ()
 			{
 				Harness = this,
 			};
