+++ tests/xharness/Properties/AssemblyInfo.cs
@@ -24,4 +24,5 @@ using System.Runtime.CompilerServices;
 
 //[assembly: AssemblyDelaySign(false)]
 //[assembly: AssemblyKeyFile("")]
+[assembly: InternalsVisibleTo("Xharness.Tests")]
