+++ tests/xharness/TestPlatform.cs
@@ -1,0 +1,23 @@
+﻿namespace Xharness
+{
+	public enum TestPlatform
+	{
+		None,
+		All,
+
+		iOS,
+		iOS_Unified,
+		iOS_Unified32,
+		iOS_Unified64,
+		iOS_TodayExtension64,
+		tvOS,
+		watchOS,
+		watchOS_32,
+		watchOS_64_32,
+
+		Mac,
+		Mac_Modern,
+		Mac_Full,
+		Mac_System,
+	}
+}
