+++ tests/xharness/Xharness.Tests/packages.config
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Castle.Core" version="4.4.0" targetFramework="net48" />
-  <package id="Moq" version="4.13.1" targetFramework="net48" />
-  <package id="NUnit" version="3.12.0" targetFramework="net47" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net47" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net47" />
+  <package id="Castle.Core" version="4.4.0" targetFramework="net472" />
+  <package id="Moq" version="4.13.1" targetFramework="net472" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net472" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net472" />
 </packages>
